% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterPerformanceResult.R
\name{as.data.frame.OuterPerformanceResult}
\alias{as.data.frame.OuterPerformanceResult}
\title{Convert OuterPerformanceResult to data.frame}
\usage{
\method{as.data.frame}{OuterPerformanceResult}(x, row.names, optional, ...)
}
\arguments{
\item{x}{[\code{\link{OuterPerformanceResult}}]}

\item{row.names}{[character]\cr
Row names for result.
Default is none.}

\item{optional}{(any)\cr
Currently ignored.}

\item{...}{[\code{...}]
Arguments passed do \code{\link{as.data.frame}}}
}
\value{
[\code{data.frame}]
}
\description{
Basically just extracts the data as a data.table
}
