% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOuterPerformances.R
\name{calcOuterPerformances}
\alias{calcOuterPerformances}
\alias{OuterPerformanceResult}
\title{Calculates Outer Performances}
\usage{
calcOuterPerformances(tuning.resampled, ...)
}
\arguments{
\item{tuning.resampled}{[\code{\link[mlr]{ResampleResult}} | \code{\link{ResampleOverfitResult}}]
Make sure to run \code{resample(..., extract = getTuneResult, keep.pred = TRUE)}.}

\item{...}{Arguments passed to specific methods}

\item{task}{[\code{\link[mlr]{Task}}]}

\item{measures}{[\code{\link[mlr]{Measure}}]}

\item{only.on.improvement}{[\code{logical(1)}]
Calculating the performances can be epensive.
Do you just want to calculate them, when a new good performance is detected in the inner performance?
Default is \code{FALSE}.}
}
\value{
[\code{OuterPerformanceResult}]
}
\description{
Calculates the performance on the outer test set for a nested resampling with a tune wrapper
}
