% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.OverfitAnalysisResult}
\alias{plot.OverfitAnalysisResult}
\title{Plots the simulated results}
\usage{
\method{plot}{OverfitAnalysisResult}(overfit.analysis)
}
\arguments{
\item{overfit.analysis}{[\code{\link{OverfitAnalysisResult}}]}
}
\description{
Plots the result of \code{\link{simulateOuterPerformance}}
}
