% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateOuterTestErrorResult.R
\name{plot.SimulateOuterTestErrorResult}
\alias{plot.SimulateOuterTestErrorResult}
\title{Plots the simulated results}
\usage{
\method{plot}{SimulateOuterTestErrorResult}(outer.performance,
  resample.overfit = NULL)
}
\arguments{
\item{outer.performance}{[\code{\link{SimulateOuterTestErrorResult}}]}

\item{resample.overfit}{[\code{\link{ResampleOverfitResult}}]
If supplied boxplots for the untuned and the tuned performances will be shown.}
}
\description{
Plots the result of \code{\link{simulateOuterTestError}}
}
