% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleOverfit.R
\name{resampleOverfit}
\alias{resampleOverfit}
\alias{ResampleOverfitResult}
\title{Resampling Generating the Necessary Data for Analysis}
\usage{
resampleOverfit(learner.tuned, task, resampling, measures = NULL, ...)
}
\arguments{
\item{learner.tuned}{[\code{\link[mlr]{Learner}}]
Learner wrapped with [\code{\link[mlr]{makeTuneWrapper}}]}

\item{task}{[\code{\link[mlr]{Task}}]}

\item{resampling}{[\code{\link[mlr]{ResampleDesc}} | \code{\link[mlr]{ResampleInstance}} ]}

\item{measures}{[\code{\link[mlr]{Measure}}]
Only the first measure will be used for analysis}
}
\value{
[\code{ResampleOverfitResult}]
}
\description{
Calls \code{\link[mlr]{resample}} with all the arguments that are nesseccary to generate the data for the overfitting analysis.
}
