% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOuterPerformance.R
\name{simulateOuterPerformance}
\alias{simulateOuterPerformance}
\alias{SimulateOuterPerformanceResult}
\title{Simulates the outer test error}
\usage{
simulateOuterPerformance(outer.performance)
}
\arguments{
\item{outer.performance}{[\code{\link{OuterPerformanceResult}}]}
}
\value{
[\code{SimulateOuterPerformanceResult}]
}
\description{
Based on the best observed point in the inner resampling the performance on the outer test set will be chosen.
}
