% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOuterTestError.R
\name{simulateOuterTestError}
\alias{simulateOuterTestError}
\alias{SimulateOuterTestErrorResult}
\title{Simulates the outer test error}
\usage{
simulateOuterTestError(outer.performance)
}
\arguments{
\item{outer.performance}{[\code{\link{OuterPerformanceResult}}]}
}
\value{
[\code{SimulateOuterTestErrorResult}]
}
\description{
Based on the best observed point in the inner resampling the performance on the outer test set will be chosen.
}
