% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterPerformanceResult.R
\name{as.data.frame.OuterPerformanceResult}
\alias{as.data.frame.OuterPerformanceResult}
\title{Convert OuterPerformanceResult to data.frame}
\usage{
\method{as.data.frame}{OuterPerformanceResult}(obj, ...)
}
\arguments{
\item{obj}{[\code{\link{OuterPerformanceResult}}]}
}
\value{
[\code{data.frame}]
}
\description{
Basically just extracts the data as a data.table
}
