% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOuterPerformances.R
\name{calcOuterPerformances}
\alias{calcOuterPerformances}
\title{Calculates Outer Performances}
\usage{
calcOuterPerformances(tuning.resampled, task, measures)
}
\arguments{
\item{tuning.resampled}{[\code{\link[mlr]{ResampleResult}}]
Make sure to run \code{resample(..., extract = getTuneResult, keep.pred = TRUE)}.}

\item{task}{[\code{\link[mlr]{Task}}]}

\item{measures}{[\code{\link[mlr]{Measure}}]}
}
\value{
\code{data.table}
}
\description{
Calculates the performance on the outer test set for a nested resampling with a tune wrapper
}
