% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.OverfitAnalysisResult}
\alias{plot.OverfitAnalysisResult}
\title{Plots the simulated results}
\usage{
\method{plot}{OverfitAnalysisResult}(x, ...)
}
\arguments{
\item{x}{[\code{\link{OverfitAnalysisResult}}]}

\item{...}{Not used.}
}
\description{
Plots the result of \code{\link{simulateOuterPerformance}}
}
