% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.SimulateOuterPerformanceResult}
\alias{plot.SimulateOuterPerformanceResult}
\title{Plots the simulated results}
\usage{
\method{plot}{SimulateOuterPerformanceResult}(outer.performance,
  resample.overfit = NULL)
}
\arguments{
\item{outer.performance}{[\code{\link{SimulateOuterPerformanceResult}}]}

\item{resample.overfit}{[\code{\link{ResampleOverfitResult}}]
If supplied boxplots for the untuned and the tuned performances will be shown.}
}
\description{
Plots the result of \code{\link{simulateOuterPerformance}}
}
